ipq40xx: Add support for D-Link DAP-2610
- One ethernet (PoE).
- Two wlan radios.
- One two colour led.
- One reset / factory default button.
- Easily accessible serial pads on pcb.
- U-Boot bootloader with tftp and "emergency web server".
- 16 MB nor flash.

Installation
============
Flash factory image from D-Link web UI. Constraints in the D-Link web UI
makes the factory image unnecessarily large. Use sysupgrade from inside
OpenWrt to reclaim some flash space.

Return to stock D-Link firmware
===============================
Partition layout is preserved, and it is possible to return to the stock
firmware simply by downloading it from D-Link and writing it to the
firmware partition.

    # mtd -r write dap2610-firmware.bin firmware

Quirks
======
To be flashable from the D-Link http server, the firmware must be larger
then 6MB, and the size in the firmware header must match the actual file
size. Also, the boot loader verifies the checksum of the firmware before
each boot, thus the jffs2 must be after the checksum covered part. This
is solved in the factory image by having the rootfs at the end of the
image (without pad-rootfs).

The sysupgrade image which does not have to be flashable from the D-Link
http server may be smaller, and the checksum in the firmware header
only covers the kernel part of the image.

Signed-off-by: Fredrik Olofsson <fredrik.olofsson@anyfinetworks.com>